+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/list/OnRealEstateClickListener.kt
@@ -1,5 +1,5 @@
 package com.openclassrooms.realestatemanager.ui.list
 
 interface OnRealEstateClickListener {
-    fun onRealEstateClick(id: Int)
+    fun onRealEstateClick(id: Long)
 }
